model-Deployment/requirements.txt
ultralytics
opencv-python
psycopg2-binary
boto3